[UPT&ADD] 增加程序启动脚本和修改重启脚本

+++ restart.sh
@@ -11,12 +11,18 @@ pid=$(pgrep myBlog)
 #
 #fi
 
-while [ $pid > /dev/null ]
-do
-    echo "myBlog pid is :$pid, it will be kill and restart soon"
+if [ -n "$pid" ]; then
+  echo "myBlog is running with pid: $pid"
+  while kill -0 $pid 2>/dev/null; do
+    echo "Killing myBlog with pid: $pid"
     kill $pid
+    sleep 1  # 添加延时，防止立即重试导致的问题
     pid=$(pgrep myBlog)
-done
+  done
+  echo "myBlog has been killed."
+else
+  echo "myBlog is not running"
+fi
 
 cd "$app_path"; rm -rf myBlog; go build -o myBlog;  ./myBlog &
